requirements.txt
tinytroupe
mcp
python-dotenv
google-generativeai
streamlit
pandas
matplotlib
fpdf2
duckduckgo-search
textblob
pydantic
